DONE: CPP04

+++ CPP03/ex01/ClapTrap.cpp
@@ -6,7 +6,7 @@
 /*   By: dim <dim@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/26 01:18:36 by dim               #+#    #+#             */
-/*   Updated: 2022/04/26 05:17:52 by dim              ###   ########.fr       */
+/*   Updated: 2022/04/26 17:20:49 by dim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,7 @@ void	ClapTrap::takeDamage(unsigned int amount)
 	}
 	std::cout << "ClapTrap < " << _name << " > has taken [ " 
 			<< _attackDamage << " ] points of damage!" <<std::endl;
-	if (_hitPoint - amount <= 0)
+	if (static_cast<int>(_hitPoint - amount) <= 0)
 	{
 		_hitPoint = 0;
 		std::cout << "              * [ 0 ] energy points left. he's dead.. *" << std::endl;
